feat: blueprints
