+++ docs/sso_authentication/user.py
@@ -1,6 +1,6 @@
 from flask_login import UserMixin
 
-from sso_authentication.db import get_db
+from .db import get_db
 
 
 class User(UserMixin):
